Changed some small things in code.
Because:
- Code cleaning

I removed some print statements and changes other small things.
